+++ goabroad-model/src/main/java/com/goabroad/model/entity/PlanMilestone.java
@@ -69,7 +69,7 @@ public class PlanMilestone extends BaseEntity {
      * 排序
      */
     @Column(name = "sort_order", nullable = false)
-    private Integer sortOrder;
+    private Short sortOrder;
     
     /**
      * 完成时间
